+++ lua/codecompanion/interactions/chat/init.lua
@@ -544,7 +544,10 @@ function Chat.new(args)
     bufnr = self.bufnr,
     messages = self.messages,
   })
-  self.tool_registry = require("codecompanion.interactions.chat.tool_registry").new({ chat = self })
+  self.tool_registry = require("codecompanion.interactions.chat.tool_registry").new({
+    chat = self,
+    ctx = self:make_system_prompt_context(),
+  })
 
   self.ui = require("codecompanion.interactions.chat.ui").new({
     adapter = self.adapter,
@@ -575,7 +578,11 @@ function Chat.new(args)
 
   if not self.hidden then
     self.close_last_chat()
-    self.ui:open():render(self.buffer_context, self.messages, { stop_context_insertion = args.stop_context_insertion })
+    self.ui:open():render(self.buffer_context, self.messages, {
+      stop_context_insertion = args.stop_context_insertion,
+      auto_submit = args.auto_submit,
+      from_prompt_library = args.from_prompt_library,
+    })
   else
     self.ui:render(self.buffer_context, self.messages, { stop_context_insertion = args.stop_context_insertion })
   end